:recycle: Update DiscreetModeToggle after code review
