+++ src/components/discreetModeToggle/discreetModeToggle.css
@@ -1,4 +1,5 @@
-@import '../../app/mixins.css';
+@import '../app/variables.css';
+@import '../app/mixins.css';
 
 .wrapper {
   width: 100%;
@@ -18,8 +19,13 @@
   justify-content: center;
   align-items: center;
   cursor: pointer;
+  user-select: none;
 
   & > span {
     margin: 0 5px;
   }
+}
+
+.rightAlignment {
+  flex-direction: row-reverse;
 }
